reorganize snake status
