+++ fighters/snake/src/status/special_n.rs
@@ -1,0 +1,9 @@
+use super::*;
+use globals::*;
+ 
+
+pub fn install() {
+    install_status_scripts!(
+        
+    );
+}
